refactor: layout header
