+++ package.json
@@ -48,7 +48,7 @@
     "@react-navigation/material-top-tabs": "^6.6.2",
     "@react-navigation/native": "^6.0.2",
     "@react-navigation/native-stack": "^6.1.0",
-    "@ribon.io/shared": "^1.1.17",
+    "@ribon.io/shared": "^1.1.18",
     "@sentry/react-native": "5.5.0",
     "@stripe/stripe-react-native": "^0.35.0",
     "@walletconnect/modal-react-native": "^1.0.0-rc.11",
