+++ src/components/moleculars/LayoutHeader/index.tsx
@@ -1,33 +1,18 @@
-import { useCurrentUser } from "contexts/currentUserContext";
-import { useCallback, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { Alert, Text, View } from "react-native";
 import CogIcon from "components/vectors/CogIcon";
 import { TouchableOpacity } from "react-native";
-import Modal from "react-native-modal";
 
-import { useNavigation } from "hooks/useNavigation";
-import { theme } from "@ribon.io/shared/styles";
-import Icon from "components/atomics/Icon";
 import { useTranslation } from "react-i18next";
 import { useWalletContext } from "contexts/walletContext";
 import WalletIcon from "components/vectors/WalletIcon";
 import { walletTruncate } from "lib/formatters/walletTruncate";
-import { openInWebViewer } from "lib/linkOpener";
-import { Linking, Platform } from "react-native";
-import ButtonSwitch from "components/atomics/buttons/ButtonSwitch";
-import { isNotificationsEnabled } from "lib/notifications";
-import { useFocusEffect } from "@react-navigation/native";
 import { logEvent } from "services/analytics";
-import { useSubscriptions } from "@ribon.io/shared/hooks";
-import { EXPO_PUBLIC_ZENDESK_KEY } from "utils/constants/Application";
 
-import ConfigItem from "../ConfigItem";
-import TicketModal from "./TicketModal";
-import ChangeLanguageItem from "./ChangeLanguageItem";
-import DeleteAccountModal from "./DeleteAccountModal";
-import LogoutModal from "./LogoutModal";
+import TicketModal from "./modals/TicketModal";
 import S from "./styles";
 import TicketSection from "./TicketSection";
+import ConfigMenu from "./ConfigMenu";
 
 type Props = {
   hideTicket?: boolean;
@@ -42,16 +27,7 @@ function LayoutHeader({
   });
   const [menuVisible, setMenuVisible] = useState(false);
   const [ticketModalVisible, setTicketModalVisible] = useState(false);
-  const [deleteAccountModalVisible, setDeleteAccountModalVisible] =
-    useState(false);
-  const [logoutModalVisible, setLogoutModalVisible] = useState(false);
-  const [notificationsEnabled, setNotificationsEnabled] = useState(false);
-  const { navigateTo } = useNavigation();
-  const { currentUser } = useCurrentUser();
-
   const { connectWallet, wallet, killSession } = useWalletContext();
-  const { userSubscriptions } = useSubscriptions();
-  const { subscriptions, refetch: refetchSubscription } = userSubscriptions();
 
   useEffect(() => {
     if (menuVisible) logEvent("P18_view");
@@ -69,252 +45,6 @@ function LayoutHeader({
   const toggleModal = () => {
     setMenuVisible(!menuVisible);
   };
-
-  const handleOpenSettings = () => {
-    if (Platform.OS === "ios") {
-      Linking.openURL("app-settings:");
-    } else {
-      Linking.openSettings();
-    }
-  };
-
-  const toggleDeleteAccountModal = () => {
-    toggleModal();
-    logEvent("deleteAccountBtn_click");
-
-    setTimeout(() => {
-      setDeleteAccountModalVisible(!deleteAccountModalVisible);
-    }, 800);
-  };
-
-  const toggleLogoutModal = () => {
-    toggleModal();
-    logEvent("signoutBtn_click");
-
-    setTimeout(() => {
-      setLogoutModalVisible(!logoutModalVisible);
-    }, 800);
-  };
-
-  useFocusEffect(
-    useCallback(() => {
-      isNotificationsEnabled().then((enabled) =>
-        setNotificationsEnabled(Boolean(enabled)),
-      );
-    }, []),
-  );
-
-  const renderTicketModal = () => {
-    if (hideTicket) return <View />;
-
-    return (
-      <TicketModal
-        visible={ticketModalVisible}
-        setVisible={setTicketModalVisible}
-      />
-    );
-  };
-
-  const renderDeleteAccountModal = () => (
-    <DeleteAccountModal
-      visible={deleteAccountModalVisible}
-      setVisible={setDeleteAccountModalVisible}
-    />
-  );
-
-  const renderLogoutModal = () => (
-    <LogoutModal
-      visible={logoutModalVisible}
-      setVisible={setLogoutModalVisible}
-    />
-  );
-
-  const handleMonthlyContributionClick = () => {
-    refetchSubscription();
-
-    logEvent("manageSubs_click", {
-      from: "configPage",
-    });
-
-    if (!currentUser) return navigateTo("PromotersScreen");
-
-    if (subscriptions && subscriptions?.length > 0) {
-      return navigateTo("MonthlyContributionsScreen");
-    } else {
-      return navigateTo("PromotersScreen");
-    }
-  };
-
-  useFocusEffect(
-    useCallback(() => {
-      refetchSubscription();
-    }, []),
-  );
-
-  const linkToSupport = () => {
-    const key = EXPO_PUBLIC_ZENDESK_KEY;
-    openInWebViewer(t("supportLink", { key }));
-    logEvent("supportBtn_click", {
-      from: "config_page",
-    });
-  };
-
-  const notificationsSwitch = () => (
-    <ButtonSwitch
-      leftText=""
-      rightText=""
-      onSwitch={() => handleOpenSettings()}
-      initialCheckState={notificationsEnabled || false}
-    />
-  );
-
-  const renderSignInConfigItem = () =>
-    !currentUser && (
-      <ConfigItem
-        icon={{
-          name: "account_circle",
-          type: "rounded",
-          color: theme.colors.brand.primary[600],
-          size: 24,
-        }}
-        text={t("signInOrCreateAccount")}
-        onPress={() => navigateTo("SignInScreen")}
-        cta={
-          <Icon
-            type="rounded"
-            size={20}
-            color={theme.colors.brand.primary[600]}
-            name="arrow_forward_ios"
-            onPress={() => navigateTo("SignInScreen")}
-          />
-        }
-      />
-    );
-
-  const renderLogoutConfigItem = () =>
-    currentUser && (
-      <ConfigItem
-        icon={{
-          name: "logout",
-          type: "rounded",
-          color: theme.colors.brand.primary[600],
-          size: 24,
-        }}
-        text={t("logout")}
-        onPress={toggleLogoutModal}
-        cta={
-          <Icon
-            type="rounded"
-            size={20}
-            color={theme.colors.brand.primary[600]}
-            name="arrow_forward_ios"
-            onPress={toggleLogoutModal}
-          />
-        }
-      />
-    );
-
-  const renderDeleteAccountConfigItem = () =>
-    currentUser && (
-      <ConfigItem
-        icon={{
-          name: "delete_forever",
-          type: "rounded",
-          color: theme.colors.brand.tertiary[400],
-          size: 24,
-        }}
-        text={t("deleteAccount")}
-        onPress={toggleDeleteAccountModal}
-        last={Boolean(currentUser)}
-        cta={
-          <Icon
-            type="rounded"
-            size={20}
-            color={theme.colors.brand.tertiary[400]}
-            name="arrow_forward_ios"
-          />
-        }
-      />
-    );
-  const renderConfigModal = () => (
-    <Modal
-      isVisible={menuVisible}
-      animationIn="slideInRight"
-      animationOut="slideOutRight"
-      hasBackdrop
-      backdropOpacity={0.2}
-      onBackdropPress={toggleModal}
-    >
-      <View style={S.supportContainer}>
-        {renderSignInConfigItem()}
-        <ConfigItem
-          icon={{
-            name: "notifications",
-            type: "rounded",
-            color: theme.colors.brand.primary[600],
-            size: 24,
-          }}
-          text={t("notifications")}
-          linkIcon={notificationsSwitch}
-        />
-
-        <ConfigItem
-          icon={{
-            name: "language",
-            type: "rounded",
-            color: theme.colors.brand.primary[600],
-            size: 24,
-          }}
-          text={t("language")}
-          linkIcon={ChangeLanguageItem}
-        />
-
-        <ConfigItem
-          icon={{
-            name: "volunteer_activism",
-            type: "rounded",
-            color: theme.colors.brand.primary[600],
-            size: 24,
-          }}
-          text={t("monthlyContributions")}
-          onPress={handleMonthlyContributionClick}
-          cta={
-            <Icon
-              type="rounded"
-              size={20}
-              color={theme.colors.brand.primary[600]}
-              name="arrow_forward_ios"
-              onPress={handleMonthlyContributionClick}
-            />
-          }
-        />
-
-        <ConfigItem
-          icon={{
-            name: "support_agent",
-            type: "rounded",
-            color: theme.colors.brand.primary[600],
-            size: 24,
-          }}
-          text={t("support")}
-          onPress={linkToSupport}
-          last={!currentUser}
-          cta={
-            <Icon
-              type="rounded"
-              size={20}
-              color={theme.colors.brand.primary[600]}
-              name="arrow_forward_ios"
-              onPress={linkToSupport}
-            />
-          }
-        />
-
-        {renderLogoutConfigItem()}
-        {renderDeleteAccountConfigItem()}
-      </View>
-    </Modal>
-  );
 
   return (
     <View style={S.configContainer}>
@@ -335,6 +65,13 @@ function LayoutHeader({
         </TouchableOpacity>
       )}
 
+      {!hideTicket && (
+        <TicketModal
+          visible={ticketModalVisible}
+          setVisible={setTicketModalVisible}
+        />
+      )}
+
       <TouchableOpacity
         accessibilityRole="button"
         style={S.container}
@@ -343,13 +80,7 @@ function LayoutHeader({
         <CogIcon />
       </TouchableOpacity>
 
-      {renderTicketModal()}
-
-      {renderConfigModal()}
-
-      {renderLogoutModal()}
-
-      {renderDeleteAccountModal()}
+      <ConfigMenu toggleModal={toggleModal} menuVisible={menuVisible} />
     </View>
   );
 }
