+++ src/lib/localStorage/constants.ts
@@ -8,5 +8,8 @@ export const COMES_FROM_TREASURE = "COMES_FROM_TREASURE";
 export const SUBGRAPH_URL = "SUBGRAPH_URL";
 export const CHAIN_ID = "CHAIN_ID";
 
+export const DONATION_TOAST_SEEN_AT_KEY = "DONATION_TOAST_SEEN_AT_KEY";
+export const DONATION_TOAST_INTEGRATION = "DONATION_TOAST_INTEGRATION";
+
 export const ACCESS_TOKEN_KEY = "ACCESS_TOKEN_KEY";
 export const REFRESH_TOKEN_KEY = "REFRESH_TOKEN_KEY";
